backend/requirements-test.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.13.0
celery[redis]==5.4.0
redis==5.2.0
httpx==0.27.0
pyjwt==2.9.0
bcrypt==4.2.0
cryptography==43.0.0
pydantic-settings==2.5.0
python-multipart==0.0.12
lxml==5.3.0
tenacity==9.0.0
ldap3>=2.9.1
openpyxl==3.1.5
pytest==8.3.0
pytest-asyncio==0.24.0
httpx[http2]==0.27.0
aiosqlite==0.20.0
aiosmtplib==3.0.2
dnspython==2.7.0
kubernetes==31.0.0
PyYAML==6.0.2
